[fix] ztf under gui/resources/bin has no exec permission
When cross-compiling non-windows package on windows, the ztf server
under gui/resources/bin will not have execution permission thus it
will fail to start with return code of 126.
When this happens, ztf GUI will report 'server not available'.
This fix is to make sure the ztf always has execution permission.
[Note] It's not recommended that user cross-compile non-windows
packages on windows.